Refactor to move files to other moduls with a main.py access point
