+++ src/types/file_details.py
@@ -1,0 +1,7 @@
+from typing import TypedDict
+
+class FileDetails(TypedDict):
+    display_text: str
+    file_name: str
+    url: str
+    downloaded: bool
